feat: add device repository

+++ internal/core/port/device.go
@@ -9,7 +9,7 @@ import (
 type DeviceRepository interface {
 	CreateDevice(ctx context.Context, device *domain.Device) (*domain.Device, error)
 	GetDeviceByID(ctx context.Context, id string) (*domain.Device, error)
-	ListDevices(ctx context.Context, page, limit int64) ([]domain.Device, error)
+	ListDevices(ctx context.Context, page, limit int) ([]domain.Device, error)
 	UpdateDevice(ctx context.Context, device *domain.Device) (*domain.Device, error)
 	DeleteDevice(ctx context.Context, id string) error
 }
@@ -17,7 +17,7 @@ type DeviceRepository interface {
 type DeviceService interface {
 	CreateDevice(ctx context.Context, device *domain.Device) (*domain.Device, error)
 	GetDevice(ctx context.Context, id string) (*domain.Device, error)
-	ListDevices(ctx context.Context, page, limit int64) ([]domain.Device, error)
+	ListDevices(ctx context.Context, page, limit int) ([]domain.Device, error)
 	UpdateDevice(ctx context.Context, device *domain.Device) (*domain.Device, error)
 	DeleteDevice(ctx context.Context, id string) error
 }
